+++ swiftui_project/swiftui_project/ScreenWithGenres/CarouselOfGenres/GenreButton/GenreButton.swift
@@ -3,8 +3,6 @@ import SwiftUI
 
 struct GenreButton: View {
     @State private var isSelected: Bool = false
-    @State private var id: String
-    private var viewModel = GenreButtonViewModel()
 
     let title: String
     let emoji: String
@@ -17,14 +15,12 @@ struct GenreButton: View {
         self.title = title
         self.emoji = emoji
         self.backgroundColor = backgroundColor
-        self._id = State(initialValue: id)
     }
     
     var body: some View {
         VStack(spacing: UISize.size8) {
             Button(emoji) {
                 isSelected.toggle()
-                viewModel.isSelectedToogle(isSelected: isSelected, id: id)
             }
             .textStyle(
                 size: UISize.size24,
