--- swiftui_project/swiftui_project/ScreenWithGenres/CarouselOfGenres/GenreButton/ViewModel/GenreButtonViewModelProtocol.swift
@@ -1,3 +1,0 @@
-protocol GenreButtonViewModelProtocol {
-    func isSelectedToogle(isSelected: Bool, id: String)
-}
