+++ swiftui_project/swiftui_project/ScreenWithGenres/CarouselOfGenres/ViewModel/GenresViewWithScrollViewModel.swift
@@ -5,10 +5,32 @@ class GenresViewWithScrollViewModel: ObservableObject {
     private var subscriber: AnyCancellable?
     private let dataManager: DataManager
     @Published var fetchedGenres: [GenreResponseModel] = []
+    @Published var genres: [String] = []
     
     init() {
         dataManager = DataManager()
-
+    }
+    
+    private  func addGenre(_ id: String) {
+        if !genres.contains(id) {
+            genres.append(id)
+        }
+    }
+    
+    private func removeGenre(_ id: String) {
+        if genres.contains(id) {
+            genres.removeAll(where: { $0 == id })
+        }
+    }
+    
+    func isSelectedToogle(isSelected: Bool, id: String) {
+        if isSelected {
+            addGenre(id)
+            print(genres)
+        } else {
+            removeGenre(id)
+            print(genres)
+        }
     }
 }
 
@@ -33,3 +55,4 @@ extension GenresViewWithScrollViewModel: GenresViewWithScrollViewModelProtocol {
             })
     }
 }
+
